Move Reads to separate file
